add tests

+++ test/cython_ext/filter_test.py
@@ -133,7 +133,7 @@ class FilterTest(TestHelper):
         f_end = np.asarray([1, 4], dtype=np.int32)
         start_index = np.asarray([0, 10, 20, 30, 40], dtype=np.int32)
         double_array = np.arange(0, 0.45, step=0.01, dtype=np.double)
-        periods = np.asarray([0, 1, 1, 0, 1])
+        periods = np.asarray([0, 1, 1, 0, 1], dtype=np.int32)
         result = good_periods(f_start,
                               f_end,
                               start_index,
@@ -154,7 +154,7 @@ class FilterTest(TestHelper):
         f_end = np.asarray([5, 7], dtype=np.int32)
         start_index = np.asarray([0, 10, 20, 30, 40], dtype=np.int32)
         double_array = np.arange(0, 0.45, step=0.01, dtype=np.double)
-        periods = np.asarray([0, 1, 1, 0, 1])
+        periods = np.asarray([0, 1, 1, 0, 1], dtype=np.int32)
         result = good_periods(f_start,
                               f_end,
                               start_index,
